app/requirements.txt
astroid==2.4.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
dominate==2.5.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
gcloud==0.18.3
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
isort==5.6.2
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
networkx==2.5
numpy==1.19.2
oauth2client==4.1.3
packaging==20.4
pandas==1.1.3
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
rsa==4.6
sip==5.4.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.11
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
